+++ sregistry/main/google_storage/__init__.py
@@ -31,7 +31,7 @@ from .build import (
     get_ipaddress,
     get_instances 
 )
-from .delete import ( delete, destroy )
+from .delete import delete
 from .logs import ( logs, list_logs, print_log )
 from .pull import pull
 from .push import ( push, upload )
@@ -145,7 +145,6 @@ Client.pull = pull
 Client.push = push
 Client._upload = upload
 Client.delete = delete
-Client.destroy = destroy
 
 # Build functions
 Client.build = build