+++ sregistry/main/s3/delete.py
@@ -11,9 +11,9 @@ Written by Tom Schoonjans (Tom.Schoonjans@diamond.ac.uk)
 '''
 
 from sregistry.logger import bot
-from sregistry.utils import ( parse_image_name, remove_uri, confirm_delete )
+from sregistry.utils import (parse_image_name, remove_uri, confirm_delete)
 
-def remove(self, image, force=False):
+def delete(self, image, force=False):
     '''delete an image from an S3 bucket'''
 
     q = parse_image_name(remove_uri(image))
@@ -24,7 +24,7 @@ def remove(self, image, force=False):
         _object = self.bucket.Object(uri)
         _object.load() # this throws an exception if the object does not exist! -> if delete() fails no exception is thrown...
 
-        if confirm_delete(force, uri) is True:
+        if confirm_delete(uri, force) is True:
             _object.delete()
         else:
             bot.info("Delete cancelled.")